Make OSS licenses plugin safe to any application order (#321)
* Simplify LicensesCleanUpTask

Delete the entire directory instead of individual files.
This is in prepration to moving LicensesTask to a single output
directory to make it compatible with the new AGP variants API.

* Move to a single output directory for LicensesTask

This is in preparation to use the new AGP variants API to register
resource creation.

* Move to the new AGP variants API

The old API is getting removed with AGP 9.0.0.

This bumps the minimum required version of AGP to 8.2 and moves
to the new variants API.

* Make OSS licenses plugin reactive

Prior to this change there was an implicit order of application of
plugins with com.android.application having to be applied before
com.google.android.gms.oss-licenses-plugin. This change makes licenses
plugin apply all of the logic despite the order of application.

---------

Co-authored-by: Vinay Guthal <vguthal@google.com>